platemate-backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
python-dotenv
openai